+++ src/main/java/org/spin/grpc/service/PointOfSalesForm.java
@@ -95,17 +95,17 @@ import org.compiere.util.MimeType;
 import org.compiere.util.TimeUtil;
 import org.compiere.util.Trx;
 import org.compiere.util.Util;
-import org.spin.backend.grpc.common.Currency;
 import org.spin.backend.grpc.common.ProcessLog;
-import org.spin.backend.grpc.common.ProductPrice;
 import org.spin.backend.grpc.common.RunBusinessProcessRequest;
+import org.spin.backend.grpc.core_functionality.Currency;
+import org.spin.backend.grpc.core_functionality.ProductPrice;
 import org.spin.backend.grpc.pos.*;
 import org.spin.backend.grpc.pos.StoreGrpc.StoreImplBase;
 import org.spin.base.util.ConvertUtil;
 import org.spin.base.util.DocumentUtil;
 import org.spin.base.util.FileUtil;
 import org.spin.base.util.RecordUtil;
-import org.spin.base.util.convert.ConvertCommon;
+import org.spin.grpc.service.core_functionality.CoreFunctionalityConvert;
 import org.spin.pos.service.POSLogic;
 import org.spin.pos.service.bank.BankManagement;
 import org.spin.pos.service.cash.CashManagement;
@@ -1492,7 +1492,7 @@ public class PointOfSalesForm extends StoreImplBase {
 						)
 					)
 					.setCurrency(
-						ConvertCommon.convertCurrency(
+						CoreFunctionalityConvert.convertCurrency(
 							rs.getInt("C_Currency_ID")
 						)
 					)
@@ -1801,7 +1801,7 @@ public class PointOfSalesForm extends StoreImplBase {
 						)
 					)
 					.setCurrency(
-						ConvertCommon.convertCurrency(
+						CoreFunctionalityConvert.convertCurrency(
 							rs.getInt("C_Currency_ID")
 						)
 					)
@@ -2059,7 +2059,7 @@ public class PointOfSalesForm extends StoreImplBase {
 					)
 				)
 				.setDocumentType(
-					ConvertCommon.convertDocumentType(
+					CoreFunctionalityConvert.convertDocumentType(
 						bankStatement.getC_DocType_ID()
 					)
 				)
@@ -3178,14 +3178,14 @@ public class PointOfSalesForm extends StoreImplBase {
 			;
 			if(availablePaymentMethod.get_ValueAsInt("RefundReferenceCurrency_ID") > 0) {
 				tenderTypeValue.setRefundReferenceCurrency(
-					ConvertCommon.convertCurrency(
+					CoreFunctionalityConvert.convertCurrency(
 						availablePaymentMethod.get_ValueAsInt("RefundReferenceCurrency_ID")
 					)
 				);
 			}
 			if(availablePaymentMethod.get_ValueAsInt("ReferenceCurrency_ID") > 0) {
 				tenderTypeValue.setReferenceCurrency(
-					ConvertCommon.convertCurrency(
+					CoreFunctionalityConvert.convertCurrency(
 						availablePaymentMethod.get_ValueAsInt("ReferenceCurrency_ID")
 					)
 				);
@@ -3292,7 +3292,7 @@ public class PointOfSalesForm extends StoreImplBase {
 		query.setLimit(limit, offset)
 			.getIDsAsList()
 			.forEach(currencyId -> {
-				Currency.Builder currencyBuilder = ConvertCommon.convertCurrency(currencyId);
+				Currency.Builder currencyBuilder = CoreFunctionalityConvert.convertCurrency(currencyId);
 				builder.addCurrencies(currencyBuilder);
 			});
 		//	
@@ -4911,7 +4911,11 @@ public class PointOfSalesForm extends StoreImplBase {
 			)
 			.setIsModifyPrice(pos.isModifyPrice())
 			.setIsPosRequiredPin(pos.isPOSRequiredPIN())
-			.setSalesRepresentative(ConvertUtil.convertSalesRepresentative(MUser.get(pos.getCtx(), pos.getSalesRep_ID())))
+			.setSalesRepresentative(
+				CoreFunctionalityConvert.convertSalesRepresentative(
+					MUser.get(pos.getCtx(), pos.getSalesRep_ID())
+				)
+			)
 			.setTemplateCustomer(
 				POSConvertUtil.convertCustomer(
 					pos.getBPartner()
@@ -5003,7 +5007,7 @@ public class PointOfSalesForm extends StoreImplBase {
 
 		if(pos.get_ValueAsInt(ColumnsAdded.COLUMNNAME_RefundReferenceCurrency_ID) > 0) {
 			builder.setRefundReferenceCurrency(
-				ConvertCommon.convertCurrency(
+				CoreFunctionalityConvert.convertCurrency(
 					pos.get_ValueAsInt(ColumnsAdded.COLUMNNAME_RefundReferenceCurrency_ID)
 				)
 			);
@@ -5011,7 +5015,9 @@ public class PointOfSalesForm extends StoreImplBase {
 		//	Set Price List and currency
 		if(pos.getM_PriceList_ID() != 0) {
 			MPriceList priceList = MPriceList.get(Env.getCtx(), pos.getM_PriceList_ID(), null);
-			builder.setPriceList(ConvertUtil.convertPriceList(priceList));
+			builder.setPriceList(
+				CoreFunctionalityConvert.convertPriceList(priceList)
+			);
 		}
 		//	Bank Account
 		if(pos.getC_BankAccount_ID() != 0) {
@@ -5019,13 +5025,13 @@ public class PointOfSalesForm extends StoreImplBase {
 			builder.setDefaultOpeningChargeId(cashAccount.get_ValueAsInt("DefaultOpeningCharge_ID"))
 				.setDefaultWithdrawalChargeId(cashAccount.get_ValueAsInt("DefaultWithdrawalCharge_ID"))
 				.setCashBankAccount(
-					ConvertCommon.convertBankAccount(cashAccount)
+					CoreFunctionalityConvert.convertBankAccount(cashAccount)
 				);
 		}
 		//	Bank Account to transfer
 		if(pos.getCashTransferBankAccount_ID() != 0) {
 			builder.setCashBankAccount(
-				ConvertCommon.convertBankAccount(
+				CoreFunctionalityConvert.convertBankAccount(
 					pos.getCashTransferBankAccount_ID()
 				)
 			);
@@ -5033,12 +5039,14 @@ public class PointOfSalesForm extends StoreImplBase {
 		//	Warehouse
 		if(pos.getM_Warehouse_ID() > 0) {
 			MWarehouse warehouse = MWarehouse.get(Env.getCtx(), pos.getM_Warehouse_ID());
-			builder.setWarehouse(ConvertUtil.convertWarehouse(warehouse));
+			builder.setWarehouse(
+				CoreFunctionalityConvert.convertWarehouse(warehouse)
+			);
 		}
 		//	Price List
 		if(pos.get_ValueAsInt(ColumnsAdded.COLUMNNAME_DisplayCurrency_ID) > 0) {
 			builder.setDisplayCurrency(
-				ConvertCommon.convertCurrency(
+				CoreFunctionalityConvert.convertCurrency(
 					pos.get_ValueAsInt(ColumnsAdded.COLUMNNAME_DisplayCurrency_ID)
 				)
 			);
@@ -5046,7 +5054,7 @@ public class PointOfSalesForm extends StoreImplBase {
 		//	Document Type
 		if(pos.getC_DocType_ID() > 0) {
 			builder.setDocumentType(
-				ConvertCommon.convertDocumentType(
+				CoreFunctionalityConvert.convertDocumentType(
 					pos.getC_DocType_ID()
 				)
 			);
@@ -5054,7 +5062,7 @@ public class PointOfSalesForm extends StoreImplBase {
 		//	Return Document Type
 		if(pos.get_ValueAsInt("C_DocTypeRMA_ID") > 0) {
 			builder.setReturnDocumentType(
-				ConvertCommon.convertDocumentType(
+				CoreFunctionalityConvert.convertDocumentType(
 					pos.get_ValueAsInt("C_DocTypeRMA_ID")
 				)
 			);
@@ -5524,7 +5532,9 @@ public class PointOfSalesForm extends StoreImplBase {
 		MProductPricing productPricing = new MProductPricing(product.getM_Product_ID(), businessPartnerId, priceQuantity, true, null);
 		productPricing.setM_PriceList_ID(priceList.getM_PriceList_ID());
 		productPricing.setPriceDate(validFrom);
-		builder.setProduct(ConvertUtil.convertProduct(product));
+		builder.setProduct(
+			CoreFunctionalityConvert.convertProduct(product)
+		);
 		int taxCategoryId = product.getC_TaxCategory_ID();
 		Optional<MTax> optionalTax = Arrays.asList(MTax.getAll(Env.getCtx()))
 		.stream()
@@ -5536,11 +5546,15 @@ public class PointOfSalesForm extends StoreImplBase {
 		.findFirst();
 		//	Validate
 		if(optionalTax.isPresent()) {
-			builder.setTaxRate(ConvertUtil.convertTaxRate(optionalTax.get()));
+			builder.setTaxRate(
+				CoreFunctionalityConvert.convertTaxRate(
+					optionalTax.get()
+				)
+			);
 		}
 		//	Set currency
 		builder.setCurrency(
-			ConvertCommon.convertCurrency(
+			CoreFunctionalityConvert.convertCurrency(
 				priceList.getC_Currency_ID()
 			)
 		);
@@ -5584,7 +5598,7 @@ public class PointOfSalesForm extends StoreImplBase {
 			//	Get from schema
 			if(displayCurrencyId > 0) {
 				builder.setDisplayCurrency(
-					ConvertCommon.convertCurrency(
+					CoreFunctionalityConvert.convertCurrency(
 						displayCurrencyId
 					)
 				);
@@ -5623,7 +5637,7 @@ public class PointOfSalesForm extends StoreImplBase {
 								)
 							)
 							.setConversionRate(
-								ConvertCommon.convertConversionRate(
+								CoreFunctionalityConvert.convertConversionRate(
 									conversionRate
 								)
 							)
@@ -6082,7 +6096,7 @@ public class PointOfSalesForm extends StoreImplBase {
 					)
 					.setIsPosRequiredPin(availableCash.get_ValueAsBoolean(I_C_POS.COLUMNNAME_IsPOSRequiredPIN))
 					.setBankAccount(
-						ConvertCommon.convertBankAccount(
+						CoreFunctionalityConvert.convertBankAccount(
 							bankAccount
 						)
 					)
