+++ src/main/java/org/spin/grpc/service/TimeControl.java
@@ -43,7 +43,7 @@ import org.spin.backend.grpc.time_control.ResourceAssignment;
 import org.spin.backend.grpc.time_control.ResourceType;
 import org.spin.backend.grpc.time_control.TimeControlGrpc.TimeControlImplBase;
 import org.spin.backend.grpc.time_control.UpdateResourceAssignmentRequest;
-import org.spin.base.util.ConvertUtil;
+import org.spin.grpc.service.core_functionality.CoreFunctionalityConvert;
 import org.spin.service.grpc.authentication.SessionManager;
 import org.spin.service.grpc.util.db.LimitUtil;
 import org.spin.service.grpc.util.value.NumberManager;
@@ -91,7 +91,9 @@ public class TimeControl extends TimeControlImplBase {
 		;
 		
 		MUOM unitOfMeasure = MUOM.get(Env.getCtx(), resourceType.getC_UOM_ID());
-		builder.setUnitOfMeasure(ConvertUtil.convertUnitOfMeasure(unitOfMeasure));
+		builder.setUnitOfMeasure(
+			CoreFunctionalityConvert.convertUnitOfMeasure(unitOfMeasure)
+		);
 		
 		return builder;
 	}
