+++ src/main/java/org/spin/grpc/service/core_functionality/CoreFunctionality.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License                *
  * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
  ************************************************************************************/
-package org.spin.grpc.service;
+package org.spin.grpc.service.core_functionality;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
@@ -22,10 +22,8 @@ import java.util.Optional;
 import java.util.Properties;
 
 import org.adempiere.core.domains.models.I_AD_Language;
-import org.adempiere.core.domains.models.I_AD_Org;
 import org.adempiere.core.domains.models.I_C_BPartner;
 import org.adempiere.core.domains.models.I_C_Country;
-import org.adempiere.core.domains.models.I_M_Warehouse;
 import org.adempiere.exceptions.AdempiereException;
 import org.compiere.model.MBPartner;
 import org.compiere.model.MBPartnerLocation;
@@ -33,12 +31,10 @@ import org.compiere.model.MConversionRate;
 import org.compiere.model.MCountry;
 import org.compiere.model.MLanguage;
 import org.compiere.model.MLocation;
-import org.compiere.model.MOrg;
 import org.compiere.model.MRole;
 import org.compiere.model.MSystem;
 import org.compiere.model.MUOMConversion;
 import org.compiere.model.MUser;
-import org.compiere.model.MWarehouse;
 import org.compiere.model.Query;
 import org.compiere.util.CCache;
 import org.compiere.util.CLogger;
@@ -47,33 +43,28 @@ import org.compiere.util.Env;
 import org.compiere.util.TimeUtil;
 import org.compiere.util.Trx;
 import org.compiere.util.Util;
-import org.spin.backend.grpc.common.BusinessPartner;
-import org.spin.backend.grpc.common.ConversionRate;
-import org.spin.backend.grpc.common.CoreFunctionalityGrpc.CoreFunctionalityImplBase;
-import org.spin.backend.grpc.common.Country;
-import org.spin.backend.grpc.common.CreateBusinessPartnerRequest;
-import org.spin.backend.grpc.common.GetBusinessPartnerRequest;
-import org.spin.backend.grpc.common.GetConversionRateRequest;
-import org.spin.backend.grpc.common.GetCountryRequest;
-import org.spin.backend.grpc.common.GetSystemInfoRequest;
-import org.spin.backend.grpc.common.ListBusinessPartnersRequest;
-import org.spin.backend.grpc.common.ListBusinessPartnersResponse;
-import org.spin.backend.grpc.common.ListLanguagesRequest;
-import org.spin.backend.grpc.common.ListLanguagesResponse;
-import org.spin.backend.grpc.common.ListOrganizationsRequest;
-import org.spin.backend.grpc.common.ListOrganizationsResponse;
-import org.spin.backend.grpc.common.ListProductConversionRequest;
-import org.spin.backend.grpc.common.ListProductConversionResponse;
-import org.spin.backend.grpc.common.ListWarehousesRequest;
-import org.spin.backend.grpc.common.ListWarehousesResponse;
-import org.spin.backend.grpc.common.ProductConversion;
-import org.spin.backend.grpc.common.SystemInfo;
+import org.spin.backend.grpc.core_functionality.BusinessPartner;
+import org.spin.backend.grpc.core_functionality.ConversionRate;
+import org.spin.backend.grpc.core_functionality.Country;
+import org.spin.backend.grpc.core_functionality.CreateBusinessPartnerRequest;
+import org.spin.backend.grpc.core_functionality.GetBusinessPartnerRequest;
+import org.spin.backend.grpc.core_functionality.GetConversionRateRequest;
+import org.spin.backend.grpc.core_functionality.GetCountryRequest;
+import org.spin.backend.grpc.core_functionality.GetSystemInfoRequest;
+import org.spin.backend.grpc.core_functionality.Language;
+import org.spin.backend.grpc.core_functionality.ListBusinessPartnersRequest;
+import org.spin.backend.grpc.core_functionality.ListBusinessPartnersResponse;
+import org.spin.backend.grpc.core_functionality.ListLanguagesRequest;
+import org.spin.backend.grpc.core_functionality.ListLanguagesResponse;
+import org.spin.backend.grpc.core_functionality.ListProductConversionRequest;
+import org.spin.backend.grpc.core_functionality.ListProductConversionResponse;
+import org.spin.backend.grpc.core_functionality.ProductConversion;
+import org.spin.backend.grpc.core_functionality.SystemInfo;
+import org.spin.backend.grpc.core_functionality.CoreFunctionalityGrpc.CoreFunctionalityImplBase;
 import org.spin.base.Version;
 import org.spin.base.db.WhereClauseUtil;
 import org.spin.base.util.ContextManager;
-import org.spin.base.util.ConvertUtil;
 import org.spin.base.util.RecordUtil;
-import org.spin.base.util.convert.ConvertCommon;
 import org.spin.service.grpc.authentication.SessionManager;
 import org.spin.service.grpc.util.db.LimitUtil;
 import org.spin.service.grpc.util.value.TimeManager;
@@ -186,6 +177,36 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 		}
 	}
 
+	/**
+	 * Convert a Country
+	 * @param context
+	 * @param request
+	 * @return
+	 */
+	private Country.Builder getCountry(GetCountryRequest request) {
+		String key = null;
+		MCountry country = null;
+		if(request.getId() <= 0) {
+			country = ContextManager.getDefaultCountry();
+		}
+		int id = request.getId();
+		if(id > 0
+				&& country == null) {
+			key = "ID:|" + request.getId();
+			country = countryCache.put(key, country);
+			if(country == null) {
+				country = MCountry.get(Env.getCtx(), request.getId());
+			}
+		}
+		if(!Util.isEmpty(key)
+				&& country != null) {
+			countryCache.put(key, country);
+		}
+		//	Return
+		return CoreFunctionalityConvert.convertCountry(Env.getCtx(), country);
+	}
+
+
 
 	/**
 	 * TODO: Duplicated with Security service
@@ -208,7 +229,105 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 			);
 		}
 	}
-	
+
+	/**
+	 * Convert languages to gRPC
+	 * @param request
+	 * @return
+	 */
+	private ListLanguagesResponse.Builder convertLanguagesList(ListLanguagesRequest request) {
+		ListLanguagesResponse.Builder builder = ListLanguagesResponse.newBuilder();
+		new Query(
+			Env.getCtx(),
+			I_AD_Language.Table_Name,
+			"(IsSystemLanguage=? OR IsBaseLanguage=?)",
+			null
+		)
+			.setParameters(true, true)
+			.setOnlyActiveRecords(true)
+			.<MLanguage>list()
+			.parallelStream()
+			.forEach(language -> {
+				Language.Builder languaBuilder = convertLanguage(language);
+				builder.addLanguages(
+					languaBuilder
+				);
+			})
+		;
+		//	Return
+		return builder;
+	}
+
+	/**
+	 * Convert Language to gRPC
+	 * @param language
+	 * @return
+	 */
+	private static Language.Builder convertLanguage(MLanguage language) {
+		Language.Builder languaBuilder = Language.newBuilder();
+		if (language == null || language.getAD_Language_ID() <= 0) {
+			return languaBuilder;
+		}
+
+		String datePattern = language.getDatePattern();
+		String timePattern = language.getTimePattern();
+		if(Util.isEmpty(datePattern, true)) {
+			org.compiere.util.Language staticLanguage = org.compiere.util.Language.getLanguage(language.getAD_Language());
+			if(staticLanguage != null) {
+				datePattern = staticLanguage.getDateFormat().toPattern();
+			}
+			//	Validate
+			if(Util.isEmpty(datePattern, true)) {
+				datePattern = language.getDateFormat().toPattern();
+			}
+		}
+		if(Util.isEmpty(timePattern, true)) {
+			org.compiere.util.Language staticLanguage = org.compiere.util.Language.getLanguage(language.getAD_Language());
+			if(staticLanguage != null) {
+				timePattern = staticLanguage.getTimeFormat().toPattern();
+			}
+		}
+		return languaBuilder
+			.setLanguage(
+				ValueManager.validateNull(
+					language.getAD_Language()
+				)
+			)
+			.setCountryCode(
+				ValueManager.validateNull(
+					language.getCountryCode()
+				)
+			)
+			.setLanguageIso(
+				ValueManager.validateNull(
+					language.getLanguageISO()
+				)
+			)
+			.setLanguageName(
+				ValueManager.validateNull(
+					language.getName()
+				)
+			)
+			.setDatePattern(
+				ValueManager.validateNull(datePattern)
+			)
+			.setTimePattern(
+				ValueManager.validateNull(timePattern)
+			)
+			.setIsBaseLanguage(
+				language.isBaseLanguage()
+			)
+			.setIsSystemLanguage(
+				language.isSystemLanguage()
+			)
+			.setIsDecimalPoint(
+				language.isDecimalPoint()
+			)
+		;
+	}
+
+
+
 	@Override
 	public void listBusinessPartners(ListBusinessPartnersRequest request,
 			StreamObserver<ListBusinessPartnersResponse> responseObserver) {
@@ -276,7 +395,7 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 			if(request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
-			ConversionRate.Builder conversionRate = ConvertCommon.convertConversionRate(
+			ConversionRate.Builder conversionRate = CoreFunctionalityConvert.convertConversionRate(
 				getConversionRate(request)
 			);
 			responseObserver.onNext(conversionRate.build());
@@ -343,7 +462,7 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 
 		ListProductConversionResponse.Builder productConversionListBuilder = ListProductConversionResponse.newBuilder();
 		Arrays.asList(MUOMConversion.getProductConversions(context, productId)).forEach(conversion -> {
-			ProductConversion.Builder productConversionBuilder = ConvertUtil.convertProductConversion(conversion);
+			ProductConversion.Builder productConversionBuilder = CoreFunctionalityConvert.convertProductConversion(conversion);
 			productConversionListBuilder.addProductConversion(productConversionBuilder);
 		});
 		
@@ -476,7 +595,7 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 		query.setLimit(limit, offset)
 			.getIDsAsList()
 			.forEach(businessPartnerId -> {
-				BusinessPartner.Builder businessPartnerBuilder = ConvertCommon.convertBusinessPartner(businessPartnerId);
+				BusinessPartner.Builder businessPartnerBuilder = CoreFunctionalityConvert.convertBusinessPartner(businessPartnerId);
 				builder.addBusinessPartners(businessPartnerBuilder);
 			});
 		//	
@@ -630,7 +749,7 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 			});
 		});
 		//	Default return
-		return ConvertCommon.convertBusinessPartner(
+		return CoreFunctionalityConvert.convertBusinessPartner(
 			businessPartner
 		);
 	}
@@ -738,240 +857,9 @@ public class CoreFunctionality extends CoreFunctionalityImplBase {
 				.setOnlyActiveRecords(true)
 				.first();
 		//	Default return
-		return ConvertCommon.convertBusinessPartner(
+		return CoreFunctionalityConvert.convertBusinessPartner(
 			businessPartner
 		);
 	}
-	
-	/**
-	 * Convert a Country
-	 * @param context
-	 * @param request
-	 * @return
-	 */
-	private Country.Builder getCountry(GetCountryRequest request) {
-		String key = null;
-		MCountry country = null;
-		if(request.getId() <= 0) {
-			country = ContextManager.getDefaultCountry();
-		}
-		int id = request.getId();
-		if(id > 0
-				&& country == null) {
-			key = "ID:|" + request.getId();
-			country = countryCache.put(key, country);
-			if(country == null) {
-				country = MCountry.get(Env.getCtx(), request.getId());
-			}
-		}
-		if(!Util.isEmpty(key)
-				&& country != null) {
-			countryCache.put(key, country);
-		}
-		//	Return
-		return ConvertUtil.convertCountry(Env.getCtx(), country);
-	}
-	
-	/**
-	 * Convert languages to gRPC
-	 * TODO: Duplicated with Security service
-	 * @param request
-	 * @return
-	 */
-	private ListLanguagesResponse.Builder convertLanguagesList(ListLanguagesRequest request) {
-		ListLanguagesResponse.Builder builder = ListLanguagesResponse.newBuilder();
-		new Query(Env.getCtx(), I_AD_Language.Table_Name, "(IsSystemLanguage=? OR IsBaseLanguage=?)", null)
-			.setParameters(true, true)
-			.setOnlyActiveRecords(true)
-			.<MLanguage>list()
-			.forEach(language -> {
-				builder.addLanguages(ConvertUtil.convertLanguage(language));
-			});
-		//	Return
-		return builder;
-	}
 
-
-
-	/**
-	 * TODO: Duplicated with Security service
-	 */
-	@Override
-	public void listOrganizations(ListOrganizationsRequest request,
-			StreamObserver<ListOrganizationsResponse> responseObserver) {
-		try {
-			if (request == null) {
-				throw new AdempiereException("Object Request Null");
-			}
-			ListOrganizationsResponse.Builder organizationsList = listOrganizations(request);
-			responseObserver.onNext(organizationsList.build());
-			responseObserver.onCompleted();
-		} catch (Exception e) {
-			log.severe(e.getLocalizedMessage());
-			e.printStackTrace();
-			responseObserver.onError(Status.INTERNAL
-				.withDescription(e.getLocalizedMessage())
-				.withCause(e)
-				.asRuntimeException()
-			);
-		}
-	}
-
-	/**
-	 * Convert Organization to list
-	 * TODO: Duplicated with Security service
-	 * @param request
-	 * @return
-	 */
-	private ListOrganizationsResponse.Builder listOrganizations(ListOrganizationsRequest request) {
-		final int roleId = request.getRoleId();
-		if(roleId < 0) {
-			throw new AdempiereException("@AD_Role_ID@ @NotFound@");
-		}
-		MRole role = MRole.get(Env.getCtx(), roleId);
-
-		//	get from role access
-		if (role == null || !role.isActive()) {
-			throw new AdempiereException("@AD_Role_ID@ @NotFound@");
-		}
-
-		List<Object> parameters = new ArrayList<Object>();
-		String whereClause = "1 = 2";
-		//	get from role access
-		if (role.isAccessAllOrgs()) {
-			whereClause = "(EXISTS(SELECT 1 FROM AD_Role r "
-				+ "WHERE r.AD_Client_ID = AD_Org.AD_Client_ID "
-				+ "AND r.AD_Role_ID = ? "
-				+ "AND r.IsActive = 'Y') "
-				+ "OR AD_Org_ID = 0) "
-			;
-			parameters.add(role.getAD_Role_ID());
-		} else {
-			if(role.isUseUserOrgAccess()) {
-				whereClause = "EXISTS(SELECT 1 FROM AD_User_OrgAccess ua "
-					+ "WHERE ua.AD_Org_ID = AD_Org.AD_Org_ID "
-					+ "AND ua.AD_User_ID = ? "
-					+ "AND ua.IsActive = 'Y')";
-				parameters.add(Env.getAD_User_ID(Env.getCtx()));
-			} else {
-				whereClause = "EXISTS(SELECT 1 FROM AD_Role_OrgAccess ra "
-					+ "WHERE ra.AD_Org_ID = AD_Org.AD_Org_ID "
-					+ "AND ra.AD_Role_ID = ? "
-					+ "AND ra.IsActive = 'Y')";
-				parameters.add(role.getAD_Role_ID());
-			}
-		}
-		whereClause += " AND IsSummary = ? ";
-		parameters.add(false);
-
-		//	Get page and count
-		String nexPageToken = null;
-		int pageNumber = LimitUtil.getPageNumber(SessionManager.getSessionUuid(), request.getPageToken());
-		int limit = LimitUtil.getPageSize(request.getPageSize());
-		int offset = (pageNumber - 1) * limit;
-		Query query = new Query(
-			Env.getCtx(),
-			I_AD_Org.Table_Name,
-			whereClause,
-			null
-		)
-			.setParameters(parameters)
-			.setOnlyActiveRecords(true)
-			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
-		;
-		//	Count
-		int count = query.count();
-
-		ListOrganizationsResponse.Builder builder = ListOrganizationsResponse.newBuilder();
-		//	Get List
-		query.setLimit(limit, offset)
-			.<MOrg>list()
-			.forEach(organization -> {
-				builder.addOrganizations(ConvertUtil.convertOrganization(organization));
-			});
-		//	
-		builder.setRecordCount(count);
-		//	Set page token
-		if(count > limit) {
-			nexPageToken = LimitUtil.getPagePrefix(SessionManager.getSessionUuid()) + (pageNumber + 1);
-		}
-		//	Set netxt page
-		builder.setNextPageToken(
-			ValueManager.validateNull(nexPageToken)
-		);
-		return builder;
-	}
-
-
-
-	/**
-	 * TODO: Duplicated with Security service
-	 */
-	@Override
-	public void listWarehouses(ListWarehousesRequest request, StreamObserver<ListWarehousesResponse> responseObserver) {
-		try {
-			if (request == null) {
-				throw new AdempiereException("Object Request Null");
-			}
-			ListWarehousesResponse.Builder organizationsList = listWarehouses(request);
-			responseObserver.onNext(organizationsList.build());
-			responseObserver.onCompleted();
-		} catch (Exception e) {
-			log.severe(e.getLocalizedMessage());
-			e.printStackTrace();
-			responseObserver.onError(Status.INTERNAL
-				.withDescription(e.getLocalizedMessage())
-				.withCause(e)
-				.asRuntimeException()
-			);
-		}
-	}
-
-	/**
-	 * Convert warehouses list
-	 * TODO: Duplicated with Security service
-	 * @param request
-	 * @return
-	 */
-	private ListWarehousesResponse.Builder listWarehouses(ListWarehousesRequest request) {
-		ListWarehousesResponse.Builder builder = ListWarehousesResponse.newBuilder();
-		//	Get page and count
-		String nexPageToken = null;
-		int pageNumber = LimitUtil.getPageNumber(SessionManager.getSessionUuid(), request.getPageToken());
-		int limit = LimitUtil.getPageSize(request.getPageSize());
-		// int offset = (pageNumber - 1) * limit;
-
-		int organizationId = request.getOrganizationId();
-		final String whereClause = "AD_Org_ID = ? AND IsInTransit = ? ";
-		Query query = new Query(
-			Env.getCtx(),
-			I_M_Warehouse.Table_Name,
-			whereClause,
-			null
-		)
-			.setParameters(organizationId, false)
-			.setOnlyActiveRecords(true)
-			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
-		;
-		//	Count
-		int count = query.count();
-		//	Get List
-		// TODO: Fix .setLimit combined with .setApplyAccessFilter and with access record (ROWNUM error)
-		query //.setLimit(limit, offset)
-			.<MWarehouse>list()
-			.forEach(warehouse -> {
-				builder.addWarehouses(ConvertUtil.convertWarehouse(warehouse));
-			});
-		//	
-		builder.setRecordCount(count);
-		//	Set page token
-		if(count > limit) {
-			nexPageToken = LimitUtil.getPagePrefix(SessionManager.getSessionUuid()) + (pageNumber + 1);
-		}
-		//	Set netxt page
-		builder.setNextPageToken(
-			ValueManager.validateNull(nexPageToken)
-		);
-		return builder;
-	}
 }
