+++ src/main/proto/security.proto
@@ -21,6 +21,9 @@ option java_outer_classname = "ADempiereSecurity";
 import "google/api/annotations.proto";
 import "google/protobuf/struct.proto";
 
+// Base URL
+// /security/
+
 package security;
 
 // The greeting service definition.
@@ -73,10 +76,6 @@ service Security {
 	rpc ListWarehouses(ListWarehousesRequest) returns (ListWarehousesResponse) {
 		option (google.api.http) = { get: "/security/warehouses" };
 	}
-	//	Request Language List
-	rpc ListLanguages(ListLanguagesRequest) returns (ListLanguagesResponse) {
-		option (google.api.http) = { get: "/security/languages" };
-	}
 	// List Available Services
 	rpc ListServices(ListServicesRequest) returns (ListServicesResponse) {
 		option (google.api.http) = { get: "/security/services" };
@@ -115,36 +114,6 @@ message LogoutRequest {
 
 // Request a SessionInfo
 message SessionInfoRequest {
-}
-
-
-// Language Item
-message Language {
-	string language = 1;
-	string language_name = 2;
-	string language_iso = 3;
-	string country_code = 4;
-	bool is_base_language = 5;
-	bool is_system_language = 6;
-	bool is_decimal_point = 7;
-	string date_pattern = 8;
-	string time_pattern = 9;
-}
-
-message ListLanguagesRequest {
-	string filters = 1;
-	string sort_by = 2;
-	repeated string group_columns = 3;
-	repeated string select_columns = 4;
-	int32 page_size = 5;
-	string page_token = 6;
-	string search_value = 7;
-}
-
-message ListLanguagesResponse {
-	int64 record_count = 1;
-	repeated Language languages = 2;
-	string next_page_token = 3;
 }
